requirements.txt
deepface==0.0.75
Flask==2.2.2
Flask_Cors==3.0.10
imageio==2.22.1
joblib==1.2.0
keras==2.10.0
matplotlib==3.6.3
mtcnn==0.1.1
nltk==3.8.1
numpy==1.23.4
opencv_python==4.6.0.66
Pillow==9.4.0
scikit_image==0.19.3
scikit_learn==1.2.0
skimage==0.0
tensorflow==2.10.0
Werkzeug==2.2.2
